restructured and navbar changes based on token presence
